+++ frontend/app/page.tsx
@@ -1,38 +1,13 @@
-// Add this line to indicate this is a Client Component
-"use client";
-
 import { error } from "console";
 import Image from "next/image";
 import React, { useEffect, useState } from 'react';
 
 
-interface UserData {
-  name: string;
-  userID: string;
-}
-
-export default function Home() {
-  const [data, setData] = useState<UserData[]>([]); 
-
-  useEffect(() => {
-    fetch('http://localhost:5045/home/data')
-    .then(response => response.json())
-    .then(data =>{ console.log(data);  setData(data);})
-    .catch(error => console.error('Error fetching data: ', error));
-  }, []);
+export default function page() {
 
   return (
-    <div>
-      <h1>Data from backend</h1>
-      {data.length > 0 && (
-        <ul>
-          {data.map((item, index) => (
-            <li key={index}>
-              {item.name} at {item.userID}
-            </li>
-          ))}
-        </ul>
-      )}
+    <div className="flex h-full flex-col px-3 py-4 md:px-2">
+      <p> 开工</p>
     </div>
-  );
+    );
 }
